requirements.txt
google-generativeai
python-dotenv
chromadb
numpy
pandas
